% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.functions.R
\name{assess.parameter.importance}
\alias{assess.parameter.importance}
\title{Assess parameter importance}
\usage{
assess.parameter.importance(don, X, Y, method)
}
\arguments{
\item{don}{a data frame.}

\item{X}{a numeric vector.}

\item{Y}{a numeric vector.}

\item{method}{a character element ("kendall-partial-correlation-slow", "pearson-partial-correlation-fast", #' "pearson-partial-correlation-slow", "spearman-partial-correlation-slow", "negative-log-p-value", "t-test").}
}
\value{
a vector.
}
\description{
Assess parameter importance
}
