% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.functions.R
\name{compare.models}
\alias{compare.models}
\title{Compare results from box/compartment models}
\usage{
compare.models(
  solution1,
  solution2,
  age.suffix2 = "",
  ignore.vars = NULL,
  time.scope = c(0, Inf),
  tolerance = list(absolute = 2, relative = 1e-04)
)
}
\arguments{
\item{solution1}{a box/compartment model object.}

\item{solution2}{a box/compartment model object.}

\item{age.suffix2}{a character element.}

\item{ignore.vars}{a logical element.}

\item{time.scope}{a numeric vector.}

\item{tolerance}{a list with absolute and relative variables.}
}
\value{
a vector.
}
\description{
Compare results from box/compartment models
}
