\name{createCont}
\alias{createCont}
\title{createCont}
\description{Create Continuous Variable}
\usage{createCont(input, n)}
\arguments{
  \item{input}{the imported template file}
  \item{n}{the number of observation in the synthetic dataet}
}



\author{Levi Waldron and Ragheed Al Dulaimi}





