\name{decideFunc}
\alias{decideFunc}
\title{function to decide the type of the variable and the function could be used to create the vairable for the synthetic data}
\description{This function will simply decide what is the type of the variable, and what is the function from the package that
can be used to simulate hte variable. This will be based on the information specified in the excel template imported into R as 
a dataframe.}
\usage{decideFunc(object)}
\arguments{
  \item{object}{The dataframe from the imported list that is based on excel template file }
}

\value{a message about the type of the variable and the R function from the package that  should be used.}

\author{Levi Waldron and Ragheed Al Dulaimi}





