\name{episim-package}
\alias{episim-package}
\alias{episim}
\docType{package}
\title{
\packageTitle{episim}
}
\description{
\packageDescription{episim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{episim}
\packageIndices{episim}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{episim}

Maintainer: \packageMaintainer{episim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
