% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.sa.demo.R
\docType{data}
\name{EpiSim.demo}
\alias{EpiSim.demo}
\title{Box/compartment models results from 108 parameter sweeps.}
\format{
A list containing two data frames, outcomes.summary.df and parms.tried.df:
\describe{
\item{lambda.overwrite}{multiplier}
\item{delta.overweight}{multiplier}
...
}
}
\source{
\url{https://github.com/statcan/episim}
}
\usage{
EpiSim.demo
}
\description{
A dataset containing box/compartment model results from 108 parameter sweeps.
}
\keyword{datasets}
