% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.functions.R
\name{get.tornado.table}
\alias{get.tornado.table}
\title{Render a tornado table}
\usage{
get.tornado.table(
  outcome.variable,
  parameters = parms.tried.df,
  outcomes = outcomes.summary.df,
  method = "kendall-partial-correlation-slow"
)
}
\arguments{
\item{outcome.variable}{a character element.}

\item{parameters}{the parms.tried.df data frame.}

\item{outcomes}{the outcomes.summary.df data frame.}

\item{method}{a character element ("kendall-partial-correlation-slow", "pearson-partial-correlation-fast", #' "pearson-partial-correlation-slow", "spearman-partial-correlation-slow", "negative-log-p-value", "t-test").}
}
\value{
none.
}
\description{
Render a tornado table
}
\examples{
# Load demo data
data("EpiSim.demo")

# Define results
parameters.swept <- EpiSim.demo$parameters
outcomes.summary.df <- EpiSim.demo$results

get.tornado.table(
  outcome.variable = "maxInc",
  parameters = parameters.swept,
  outcomes = outcomes.summary.df
)

}
