\name{getFunc}
\alias{getFunc}
\title{function to decide and intitiate the function that sould be used to create the vairable for the synthetic data}
\description{This function will simply decide what is what is the function from the package that
should be used to simulate hte variable, then it will run this function to create the variable based on the inofrmation in the dataframe
imported into R  from the excel template file.}
\usage{getFunc(object)}
\arguments{
  \item{object}{The dataframe from the imported list that is based on excel template file }
}

\value{the output will be a dataframe that contain the simulated variable.}

\author{Levi Waldron and Ragheed Al Dulaimi}





