% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{overlap.length}
\alias{overlap.length}
\title{Computes the length of intersection of two intervals}
\usage{
overlap.length(L1, U1, L2, U2)
}
\arguments{
\item{L1}{a numeric vector.}

\item{U1}{a numeric vector.}

\item{L2}{a numeric vector.}

\item{U2}{a numeric vector.}
}
\value{
a vector.
}
\description{
Computes the length of intersection of two intervals
}
