% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.functions.R
\name{read.hypercube.sampling.specs}
\alias{read.hypercube.sampling.specs}
\title{Get hypercube sampling specifications}
\usage{
read.hypercube.sampling.specs(file.name, sheet)
}
\arguments{
\item{file.name}{the name of the Excel workbook where the specifications are located.}

\item{sheet}{the name of the sheet where the specifications are located.}
}
\value{
a list.
}
\description{
Get hypercube sampling specifications
}
