% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{rev_str}
\alias{rev_str}
\title{Reverse the characters of a character vector}
\usage{
rev_str(string)
}
\arguments{
\item{string}{a character vector of length 1.}
}
\value{
a character vector of length 1.
}
\description{
Reverse the characters of a character vector
}
\examples{
rev_str("Hello world!")
# [1] "!dlrow olleH"

}
