% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.functions.R
\name{save.model.in.workbook}
\alias{save.model.in.workbook}
\title{Save box/compartment model as an Excel workbook}
\usage{
save.model.in.workbook(input.info.list, file_name, map.names)
}
\arguments{
\item{input.info.list}{a list.}

\item{file_name}{a character element.}

\item{map.names}{a list of workbook sheet names.}
}
\value{
none.
}
\description{
Save box/compartment model as an Excel workbook
}
