% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{split_str}
\alias{split_str}
\title{Split a character vector of length 1 into several elements}
\usage{
split_str(string, separator = " ")
}
\arguments{
\item{string}{a character vector of length 1.}

\item{separator}{a character vector on which to split the \code{string} parameter. By default, " ".}
}
\value{
a character vector.
}
\description{
Split a character vector of length 1 into several elements
}
\examples{
split_str("Hello world!")
# [1] "Hello"  "world!"

}
