% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{unfactor}
\alias{unfactor}
\title{Coerce a factor vector to a numeric vector}
\usage{
unfactor(x)
}
\arguments{
\item{x}{a factor vector.}
}
\value{
a numeric vector.
}
\description{
Coerce a factor vector to a numeric vector
}
