% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{verbose.save}
\alias{verbose.save}
\title{Save R objects to a file}
\usage{
verbose.save(
  object,
  path.with.trailing.slash = "",
  prefix.suffix = c(prefix = "This file contains an R object called ", suffix =
    ".SavedFromR"),
  time.stamp = gsub(":", "-", Sys.time())
)
}
\arguments{
\item{object}{an object or a character element representing the R object to be saved.}

\item{path.with.trailing.slash}{set to null by default.}

\item{prefix.suffix}{a named (prefix, suffix), two-element character vector representing the prefix and suffix for the file name. By default, the prefix is "This file contains an R object called " and the suffix is ".SavedFromR".}

\item{time.stamp}{the format for the timestamp that appears in the file name. By default, the format is yyyy-mm-dd hh-mm-ss.}
}
\value{
none.
}
\description{
This function is a wrapper for the base \code{R} \code{\link[=save]{save()}} function that adds prefix/suffix information and compression to the R objects that are saved to file.
}
