\name{EpiSim-package}
\alias{EpiSim-package}
\alias{EpiSim}
\docType{package}
\title{functions for creating a simulated epidemiological data based on excel sheet with
descriptive statistics of the real data.}
\description{Contains functions for reading a template excel sheets as data frame and using the variables names and descriptive statistics to create synthetic datasets. The package include functions to create various variable types including categorical, continuous and date.}
\details{
\tabular{ll}{Package: \tab EpiSim\cr
Maintainer: \tab Levi Waldron <lwaldron.research@gmail.com>\cr
Depends: \tab R (>= 3.1.0)\cr
Imports: \tab XLConnect, XLConnectJars\cr
Suggests: \tab \cr
Author: \tab Levi Waldron and Ragheed Al Dulaimi\cr
Version: \tab 0.1.0\cr
Date: \tab 2014-12-24\cr
License: \tab GPL (>= 3.0)\cr
Title: \tab functions for creating a simulated epidemiological data based on excel sheet with
descriptive statistics of the real data.\cr
URL: \tab https://github.com/raldulai/Episim\cr
BugReports: \tab https://github.com/raldulai/Episim/issues\cr
Packaged: \tab 2014-12-30 21:13:32 UTC; raldulai\cr
NeedsCompilation: \tab no\cr
LazyData: \tab true\cr
Repository: \tab CRAN\cr
Date/Publication: \tab 2015-01-01 01:40:05\cr}
}
\author{Levi Waldron and Ragheed Al Dulaimi}

\keyword{ package }


