% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.functions.R
\name{try.various.parms.values}
\alias{try.various.parms.values}
\title{Perform a parameter sweep}
\usage{
try.various.parms.values(
  seir.object,
  parm.cloud.grid.specs,
  only.show.parms.to.try = FALSE
)
}
\arguments{
\item{seir.object}{a box/compartment model object.}

\item{parm.cloud.grid.specs}{a data frame of parameter sweep specifications.}

\item{only.show.parms.to.try}{a logical value. By default, FALSE.}
}
\value{
a list of parameter sweep inputs and results.
}
\description{
Perform a parameter sweep
}
